+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-28 02:56:18
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: b0cb5573713f4cf1ccfbee6c4d8387f4b7ff044d
+
+## [ForwardWidgets-2kuai] - 2025-08-28 02:56:18
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 6af0914a0f3822d2efd9b8fe1ad74c15e0a924a0
+
+## [ForwardWidgets-pack1r] - 2025-08-28 02:56:18
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 725eb2e27d05761ed5f98b7b67b40d66887994cf
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-27 02:56:29
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main